[PHPUnit 10] Move rules to own PHPUnit100 category
