+++ rules-tests/PHPUnit100/Rector/Class_/AddProphecyTraitRector/Fixture/skip_if_already_added.php.inc
@@ -1,0 +1,16 @@
+<?php
+
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\Class_\AddProphecyTraitRector\Fixture;
+
+use PHPUnit\Framework\TestCase;
+use Prophecy\PhpUnit\ProphecyTrait;
+
+final class SkipIfAlreadyAdded extends TestCase
+{
+    use ProphecyTrait;
+
+    public function testOne(): void
+    {
+        $prophecy = $this->prophesize(\AnInterface::class);
+    }
+}
