+++ rules-tests/PHPUnit100/Rector/Class_/StaticDataProviderClassMethodRector/Fixture/cover_no_bracket_name.php.inc
@@ -1,0 +1,45 @@
+<?php
+
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\Class_\StaticDataProviderClassMethodRector\Fixture;
+
+use PHPUnit\Framework\TestCase;
+
+final class CoverNoBracketName extends TestCase
+{
+    /**
+     * @dataProvider provideData
+     */
+    public function test()
+    {
+    }
+
+    public function provideData()
+    {
+        yield [1];
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\Class_\StaticDataProviderClassMethodRector\Fixture;
+
+use PHPUnit\Framework\TestCase;
+
+final class CoverNoBracketName extends TestCase
+{
+    /**
+     * @dataProvider provideData
+     */
+    public function test()
+    {
+    }
+
+    public static function provideData()
+    {
+        yield [1];
+    }
+}
+
+?>
