+++ rules-tests/PHPUnit100/Rector/Class_/StaticDataProviderClassMethodRector/Source/SomeClassToMock.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\Class_\StaticDataProviderClassMethodRector\Source;
+
+class SomeClassToMock {}
