+++ rules-tests/PHPUnit100/Rector/MethodCall/PropertyExistsWithoutAssertRector/Fixture/fixture3.php.inc
@@ -1,0 +1,29 @@
+<?php
+
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\MethodCall\AssertPropertyExistsRector\Fixture;
+
+final class MyTest3 extends \PHPUnit\Framework\TestCase
+{
+    public function test()
+    {
+        $this->assertObjectHasAttribute('property', $object->data);
+        $this->assertObjectNotHasAttribute('property', $object->data);
+    }
+}
+
+?>
+-----
+<?php
+
+namespace Rector\PHPUnit\Tests\PHPUnit100\Rector\MethodCall\AssertPropertyExistsRector\Fixture;
+
+final class MyTest3 extends \PHPUnit\Framework\TestCase
+{
+    public function test()
+    {
+        $this->assertTrue(property_exists($object->data, 'property'));
+        $this->assertFalse(property_exists($object->data, 'property'));
+    }
+}
+
+?>
