--- tests/Rector/Class_/AddProphecyTraitRector/Fixture/skip_if_already_added.php.inc
@@ -1,16 +1,0 @@
-<?php
-
-namespace Rector\PHPUnit\Tests\Rector\Class_\AddProphecyTraitRector\Fixture;
-
-use PHPUnit\Framework\TestCase;
-use Prophecy\PhpUnit\ProphecyTrait;
-
-final class SkipIfAlreadyAdded extends TestCase
-{
-    use ProphecyTrait;
-
-    public function testOne(): void
-    {
-        $prophecy = $this->prophesize(\AnInterface::class);
-    }
-}
