--- tests/Rector/Class_/StaticDataProviderClassMethodRector/Fixture/handle_abstract_data_provider.php.inc
@@ -1,39 +1,0 @@
-<?php
-
-namespace Rector\PHPUnit\Tests\Rector\Class_\StaticDataProviderClassMethodRector\Fixture;
-
-use PHPUnit\Framework\TestCase;
-
-abstract class HandleAbstractDataProvider extends TestCase
-{
-    /**
-     * @dataProvider provideData
-     */
-    public function test()
-    {
-    }
-
-    public abstract function provideData();
-}
-
-?>
------
-<?php
-
-namespace Rector\PHPUnit\Tests\Rector\Class_\StaticDataProviderClassMethodRector\Fixture;
-
-use PHPUnit\Framework\TestCase;
-
-abstract class HandleAbstractDataProvider extends TestCase
-{
-    /**
-     * @dataProvider provideData
-     */
-    public function test()
-    {
-    }
-
-    abstract public static function provideData();
-}
-
-?>
