--- tests/Rector/Class_/StaticDataProviderClassMethodRector/Fixture/respect_parent_abstract_class.php.inc
@@ -1,35 +1,0 @@
-<?php
-
-namespace Rector\PHPUnit\Tests\Rector\Class_\StaticDataProviderClassMethodRector\Fixture;
-
-use Rector\PHPUnit\Tests\Rector\Class_\StaticDataProviderClassMethodRector\Source\AnotherAbstractClass;
-
-final class RespectParentAbstractClass extends AnotherAbstractClass
-{
-    public function dataProvider(): array
-    {
-        return [
-            [1, 2],
-        ];
-    }
-}
-
-?>
------
-<?php
-
-namespace Rector\PHPUnit\Tests\Rector\Class_\StaticDataProviderClassMethodRector\Fixture;
-
-use Rector\PHPUnit\Tests\Rector\Class_\StaticDataProviderClassMethodRector\Source\AnotherAbstractClass;
-
-final class RespectParentAbstractClass extends AnotherAbstractClass
-{
-    public static function dataProvider(): array
-    {
-        return [
-            [1, 2],
-        ];
-    }
-}
-
-?>
