--- tests/Rector/Class_/StaticDataProviderClassMethodRector/Fixture/skip_already_static.php.inc
@@ -1,20 +1,0 @@
-<?php
-
-namespace Rector\PHPUnit\Tests\Rector\Class_\StaticDataProviderClassMethodRector\Fixture;
-
-use PHPUnit\Framework\TestCase;
-
-final class SkipAlreadyStatic extends TestCase
-{
-    /**
-     * @dataProvider provideData()
-     */
-    public function test()
-    {
-    }
-
-    public static function provideData()
-    {
-        yield [1];
-    }
-}
