--- tests/Rector/Class_/StaticDataProviderClassMethodRector/Source/SomeClassToMock.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace Rector\PHPUnit\Tests\Rector\Class_\StaticDataProviderClassMethodRector\Source;
-
-class SomeClassToMock {}
