--- tests/Rector/Class_/StaticDataProviderClassMethodRector/config/configured_rule.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-
-use Rector\PHPUnit\Rector\Class_\StaticDataProviderClassMethodRector;
-
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->import(__DIR__ . '/../../../../../config/config.php');
-
-    $rectorConfig->rule(StaticDataProviderClassMethodRector::class);
-};
