--- tests/Rector/MethodCall/PropertyExistsWithoutAssertRector/config/configured_rule.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-use Rector\Config\RectorConfig;
-use Rector\PHPUnit\Rector\MethodCall\PropertyExistsWithoutAssertRector;
-
-return static function (RectorConfig $rectorConfig): void {
-    $rectorConfig->import(__DIR__ . '/../../../../../config/config.php');
-
-    $rectorConfig->rule(PropertyExistsWithoutAssertRector::class);
-};
